Change location of DataflowGBK, delete DataflowGBKTranslation
